requirements.txt
anthropic>=0.40.0
pytrends>=4.9.0
google-api-python-client>=2.100.0
notion-client>=2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
yfinance>=0.2.36
numpy>=1.24.0
feedparser>=6.0.10
flask>=3.0.0
supabase>=2.0.0
waitress>=3.0.0
